rename 3 implementations
